Match API update / Local cfg file
Have somewhat gotten the matches API working. Still need to figure out getting and setting team data for the match, however a user can now start and stop a match via the button on the match cards. The match cards send a put request to set a matches "current" value to true or false. The HUD now sends a get request for current matches and will render based on the return. My solution needs more refining to account for things such as users trying to set multiple games live at the same time.

I have also added the gsi cfg file in the public folder so people do not have to download it separately.