Current Match Card / Layout
Added a card that pops up when a user starts a match. First version of it so there will probably be errors.

Also worked on how the admin panel looks. It should look a little more professional now, and I was even messing around with going the full material ui route. Maybe in the future.