Swap Sides Func, Theme tweak
Added the first attempt at reversing the sides of the teams on the hud. First crack at it so there will probably be issues. Can already see that you might have to click it twice to work if the hud is refreshed.

Have also realized that the HUD was passing classes to the Admin Panel, so a temporary fix has been made for that but need to look into other solutions.

Minor tweaks to the admin panel layout. Using a fixed position now for the side panel and some other minor updates to it.

Added a todo list in the app.tsx.