Bomb/Defuse Timers fix, Panel work
Fixed the bomb and defuse timers so that they are actually working now. Needed to make the countdown in Countdown.tsx be a number instead of string and it fixed the issue.

Also more work creating my own components so we can get rid of material ui.

Thanks to fyflo for bringing it to my attention