Database / Hud update
Reworked the look of the base hud observed player and matchbar. Looks a bit better in my opinion.

Also reworked match/team database. We now use the Teams interface from api.ts, which means that we needed to update teams to use _id rather than id.

Change the way we reverse sides to work more like the original hud does where we select reverse side in each map in the veto. Still working on adding winner and scores to the map series.