Major Rework
A lot has changed this update.
-Port changed to 1349 (need to update the gamestate.cfg in your game files)
-Hud is no longer built into the Client
-Hud now has its own repository:  https://github.com/JohnTimmermann/OpenHud-React-Hud
-Server hosts a fully built hud for better performance

To customize the hud, you can now clone it from the above repo and edit it there. Make sure you start the server before running npm start.

After you customize the hud to your liking, you will need to run `npm run build`. This will build a production build of the hud that we can use for an overlay and server.

To use it with everything, replace the build folder in the Server directory with the new one from your custom hud.

There is also a new Overlay: https://github.com/JohnTimmermann/OpenHud-Overlay

There is a few other things as well, and there will probably be some bugs so please report them in the discord:
https://discord.gg/HApB9HyaWM