#19 - Finishing touches to upload
Finished up the routes for player pictures/team logos for creating and updating teams with local images.

Had to switch to useing FormData instead of passing a Player / Team type to the functions. Will look more into this later for type safety.

Have to switch all instances of images to use the api urls to the location of the image instead of a website.

Huds must be updated to account for the new routes for images